Factor out connection code
