+++ mwmbl/batches.py
@@ -1,0 +1,3 @@
+"""
+Database interface for batches of crawled data.
+"""
